TASK-30 [Rename]: 오타 수정
